requirements.txt
Flask
pytesseract
Pillow
qrcode[pil]